requirements.txt
asgiref==3.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-webpack-loader==0.7.0
djangorestframework==3.12.1
geojson==2.5.0
gpx-converter==1.7.2
gpxpy==1.4.2
idna==2.10
mysql==0.0.2
mysql-connector-python==8.0.22
mysqlclient==2.0.1
numpy==1.19.4
pandas==1.1.4
protobuf==3.13.0
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
selenium==3.141.0
Shapely==1.7.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.11
webdriver-manager==3.2.2
channels~=3.0.2